[#26] Add delete button to each event card
